+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 ### Added
-- Additional multipart type, by @HardNorth
+- Additional multipart types, by @HardNorth
 ### Fixed
 - Response title format for empty phrase, by @HardNorth